+++ svelte.config.js
@@ -12,13 +12,13 @@ const config = {
     // If your environment is not supported or you settled on a specific environment, switch out the adapter.
     // See https://kit.svelte.dev/docs/adapters for more information about adapters.
     adapter: adapter({
-      pages: 'dist',
+      pages: 'dist'
 
       // NOTE: Don't enable `fallback`. Adding a fallback causes relative path to fail, causing Webflow app to not load from within a subdirectory.
-      // fallback: 'index.html' 
+      // fallback: 'index.html'
     }),
     paths: {
-      relative: true,
+      relative: true
     }
   }
 };