+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # 1.1.2
+
 - UI style updates to better match the new Webflow design
 - Fix editor height whenever iframe window resizes
 - Tab indentation in the editor
 
 # 1.1.1
+
 - Finally fix bundle 404 error - add relative path to build config
 
 # 1.1.0